Back-end/requirements.txt
# Core Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pgvector==0.3.6

# GraphQL
strawberry-graphql[fastapi]==0.283.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1

# Email
python-multipart==0.0.12
jinja2==3.1.4
aiosmtplib==3.0.2
mailtrap==2.0.1

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Utils
python-dateutil==2.9.0
httpx==0.27.0
